Split module into different files
This makes the code cleaner, easier to read and supports future changes by allowing the library to expand.
