+++ errors.py
@@ -1,0 +1,3 @@
+class NrrdError(Exception):
+    """Exceptions for Nrrd class."""
+    pass
