+++ reader.py
@@ -1,0 +1,296 @@
+import bz2
+import os
+import zlib
+
+import numpy as np
+
+from errors import NrrdError
+from parsers import *
+
+_NRRD_FIELD_PARSERS = {
+    'dimension': int,
+    'type': str,
+    'sizes': lambda fieldValue: parse_number_list(fieldValue, dtype=int),
+    'endian': str,
+    'encoding': lambda fieldValue: str(fieldValue).lower(),
+    'min': float,
+    'max': float,
+    'oldmin': float,
+    'old min': float,
+    'oldmax': float,
+    'old max': float,
+    'lineskip': int,
+    'line skip': int,
+    'byteskip': int,
+    'byte skip': int,
+    'content': str,
+    'sample units': str,
+    'datafile': str,
+    'data file': str,
+    'spacings': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'thicknesses': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'axis mins': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'axismins': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'axis maxs': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'axismaxs': lambda fieldValue: parse_number_list(fieldValue, dtype=float),
+    'centerings': lambda fieldValue: [str(x) for x in fieldValue.split()],
+    'labels': lambda fieldValue: [str(x) for x in fieldValue.split()],
+    'units': lambda fieldValue: [str(x) for x in fieldValue.split()],
+    'kinds': lambda fieldValue: [str(x) for x in fieldValue.split()],
+    'space': str,
+    'space dimension': int,
+    'space units': lambda fieldValue: [str(x) for x in fieldValue.split()],
+    'space origin': lambda fieldValue: parse_vector(fieldValue, dtype=float),
+    'space directions': lambda fieldValue: parse_optional_matrix(fieldValue),
+    'measurement frame': lambda fieldValue: parse_optional_matrix(fieldValue),
+}
+
+_NRRD_REQUIRED_FIELDS = ['dimension', 'type', 'encoding', 'sizes']
+
+_TYPEMAP_NRRD2NUMPY = {
+    'signed char': 'i1',
+    'int8': 'i1',
+    'int8_t': 'i1',
+    'uchar': 'u1',
+    'unsigned char': 'u1',
+    'uint8': 'u1',
+    'uint8_t': 'u1',
+    'short': 'i2',
+    'short int': 'i2',
+    'signed short': 'i2',
+    'signed short int': 'i2',
+    'int16': 'i2',
+    'int16_t': 'i2',
+    'ushort': 'u2',
+    'unsigned short': 'u2',
+    'unsigned short int': 'u2',
+    'uint16': 'u2',
+    'uint16_t': 'u2',
+    'int': 'i4',
+    'signed int': 'i4',
+    'int32': 'i4',
+    'int32_t': 'i4',
+    'uint': 'u4',
+    'unsigned int': 'u4',
+    'uint32': 'u4',
+    'uint32_t': 'u4',
+    'longlong': 'i8',
+    'long long': 'i8',
+    'long long int': 'i8',
+    'signed long long': 'i8',
+    'signed long long int': 'i8',
+    'int64': 'i8',
+    'int64_t': 'i8',
+    'ulonglong': 'u8',
+    'unsigned long long': 'u8',
+    'unsigned long long int': 'u8',
+    'uint64': 'u8',
+    'uint64_t': 'u8',
+    'float': 'f4',
+    'double': 'f8',
+    'block': 'V'
+}
+
+
+def _determine_dtype(fields):
+    """Determine the numpy dtype of the data."""
+    # Check whether the required fields are there
+    for field in _NRRD_REQUIRED_FIELDS:
+        if field not in fields:
+            raise NrrdError('Nrrd header misses required field: "%s".' % (field))
+
+    # Process the data type
+    np_typestring = _TYPEMAP_NRRD2NUMPY[fields['type']]
+    # Endianness is not necessary for ASCII encoding type
+    if np.dtype(np_typestring).itemsize > 1 and fields['encoding'] not in ['ascii', 'text', 'txt']:
+        if 'endian' not in fields:
+            raise NrrdError('Nrrd header misses required field: "endian".')
+        if fields['endian'] == 'big':
+            np_typestring = '>' + np_typestring
+        elif fields['endian'] == 'little':
+            np_typestring = '<' + np_typestring
+
+    return np.dtype(np_typestring)
+
+
+def _validate_magic_line(line):
+    """For NRRD files, the first four characters are always "NRRD", and
+    remaining characters give information about the file format version
+
+    >>> _validate_magic_line('NRRD0005')
+    8
+    >>> _validate_magic_line('NRRD0006')
+    Traceback (most recent call last):
+        ...
+    NrrdError: NRRD file version too new for this library.
+    >>> _validate_magic_line('NRRD')
+    Traceback (most recent call last):
+        ...
+    NrrdError: Invalid NRRD magic line: NRRD
+    """
+    if not line.startswith('NRRD'):
+        raise NrrdError('Missing magic "NRRD" word. Is this an NRRD file?')
+    try:
+        if int(line[4:]) > 5:
+            raise NrrdError('NRRD file version too new for this library.')
+    except ValueError:
+        raise NrrdError('Invalid NRRD magic line: %s' % (line,))
+    return len(line)
+
+
+def read_data(fields, filehandle, filename=None):
+    """Read the NRRD data from a file object into a numpy structure.
+
+    File handle is is assumed to point to the first byte of the data. That is,
+    in case of an attached header, assumed to point to the first byte after the
+    '\n\n' line.
+    """
+    data = np.zeros(0)
+    # Determine the data type from the fields
+    dtype = _determine_dtype(fields)
+    # determine byte skip, line skip, and data file (there are two ways to write them)
+    lineskip = fields.get('lineskip', fields.get('line skip', 0))
+    byteskip = fields.get('byteskip', fields.get('byte skip', 0))
+    datafile = fields.get('datafile', fields.get('data file', None))
+    datafilehandle = filehandle
+    if datafile is not None:
+        # If the datafile path is absolute, don't muck with it. Otherwise
+        # treat the path as relative to the directory in which the detached
+        # header is in
+        if os.path.isabs(datafile):
+            datafilename = datafile
+        else:
+            datafilename = os.path.join(os.path.dirname(filename), datafile)
+        datafilehandle = open(datafilename, 'rb')
+
+    num_pixels = np.array(fields['sizes']).prod()
+    # Seek to start of data based on lineskip/byteskip. byteskip == -1 is
+    # only valid for raw encoding and overrides any lineskip
+    if fields['encoding'] == 'raw' and byteskip == -1:
+        datafilehandle.seek(-dtype.itemsize * num_pixels, 2)
+    else:
+        for _ in range(lineskip):
+            datafilehandle.readline()
+
+    if fields['encoding'] == 'raw':
+        datafilehandle.seek(byteskip, os.SEEK_CUR)
+        data = np.fromfile(datafilehandle, dtype)
+    elif fields['encoding'] in ['ascii', 'text', 'txt']:
+        datafilehandle.seek(byteskip, os.SEEK_CUR)
+        data = np.fromfile(datafilehandle, dtype, sep=' ')
+    else:
+        # Probably the data is compressed then
+        if fields['encoding'] == 'gzip' or \
+                fields['encoding'] == 'gz':
+            decompobj = zlib.decompressobj(zlib.MAX_WBITS | 16)
+        elif fields['encoding'] == 'bzip2' or \
+                fields['encoding'] == 'bz2':
+            decompobj = bz2.BZ2Decompressor()
+        else:
+            raise NrrdError('Unsupported encoding: "%s"' % fields['encoding'])
+
+        decompressed_data = b''
+        while True:
+            chunk = datafilehandle.read(_READ_CHUNKSIZE)
+            if not chunk:
+                break
+            decompressed_data += decompobj.decompress(chunk)
+        # byteskip applies to the _decompressed_ byte stream
+        data = np.frombuffer(decompressed_data[byteskip:], dtype)
+
+    if datafilehandle:
+        datafilehandle.close()
+
+    if num_pixels != data.size:
+        raise NrrdError('ERROR: {0}-{1}={2}'.format(num_pixels, data.size, num_pixels - data.size))
+
+    # dkh : eliminated need to reverse order of dimensions. nrrd's
+    # data layout is same as what numpy calls 'Fortran' order,
+    shape_tmp = list(fields['sizes'])
+    data = np.reshape(data, tuple(shape_tmp), order='F')
+    return data
+
+
+def read_header(nrrdfile):
+    """Parse the fields in the nrrd header
+
+    nrrdfile can be any object which supports the iterator protocol and
+    returns a string each time its next() method is called — file objects and
+    list objects are both suitable. If nrrdfile is a file object, it must be
+    opened with the ‘b’ flag on platforms where that makes a difference
+    (e.g. Windows)
+
+    >>> read_header(("NRRD0005", "type: float", "dimension: 3"))
+    {u'type': 'float', u'dimension': 3, u'keyvaluepairs': {}}
+    >>> read_header(("NRRD0005", "my extra info:=my : colon-separated : values"))
+    {u'keyvaluepairs': {u'my extra info': u'my : colon-separated : values'}}
+    """
+    # Collect number of bytes in the file header (for seeking below)
+    header_size = 0
+
+    it = iter(nrrdfile)
+    magic_line = next(it)
+
+    need_decode = False
+    if hasattr(magic_line, 'decode'):
+        need_decode = True
+        magic_line = magic_line.decode('ascii', 'ignore')
+
+    header_size += _validate_magic_line(magic_line)
+
+    header = {u'keyvaluepairs': {}}
+    for raw_line in it:
+        header_size += len(raw_line)
+        if need_decode:
+            raw_line = raw_line.decode('ascii', 'ignore')
+
+        # Trailing whitespace ignored per the NRRD spec
+        line = raw_line.rstrip()
+
+        # Comments start with '#', no leading whitespace allowed
+        if line.startswith('#'):
+            continue
+        # Single blank line separates the header from the data
+        if line == '':
+            break
+
+        # Handle the <key>:=<value> lines first since <value> may contain a
+        # ': ' which messes up the <field>: <desc> parsing
+        key_value = line.split(':=', 1)
+        if len(key_value) is 2:
+            key, value = key_value
+            # TODO: escape \\ and \n ??
+            # value.replace(r'\\\\', r'\\').replace(r'\n', '\n')
+            header[u'keyvaluepairs'][key] = value
+            continue
+
+        # Handle the "<field>: <desc>" lines.
+        field_desc = line.split(': ', 1)
+        if len(field_desc) is 2:
+            field, desc = field_desc
+            # Preceeding and suffixing white space should be ignored.
+            field = field.rstrip().lstrip()
+            desc = desc.rstrip().lstrip()
+            if field not in _NRRD_FIELD_PARSERS:
+                raise NrrdError('Unexpected field in nrrd header: %s' % repr(field))
+            if field in header.keys():
+                raise NrrdError('Duplicate header field: %s' % repr(field))
+            header[field] = _NRRD_FIELD_PARSERS[field](desc)
+            continue
+
+        # Should not reach here
+        raise NrrdError('Invalid header line: %s' % repr(line))
+
+    # line reading was buffered; correct file pointer to just behind header:
+    if hasattr(nrrdfile, 'seek'):
+        nrrdfile.seek(header_size)
+
+    return header
+
+
+def read(filename):
+    """Read a nrrd file and return a tuple (data, header)."""
+    with open(filename, 'rb') as filehandle:
+        header = read_header(filehandle)
+        data = read_data(header, filehandle, filename)
+        return (data, header)
